datascience/requirements.txt
matplotlib
pandas
seaborn
jupyter
docopt
dotenv
pyyaml
awesome-slugify
flask
plotly
